refactoring and adding employeelist to app component
