+++ client/Components/App.js
@@ -3,6 +3,7 @@ import { Routes, Route } from "react-router-dom";
 import Home from "./Home/Home";
 import PayrollCalculator from "./PayrollCalculator/PayrollCalculator";
 import Header from "./Header/Header";
+import EmployeeList from "./Employee/EmployeeList";
 
 const App = () => {
   return (
@@ -11,6 +12,7 @@ const App = () => {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/calculator" element={<PayrollCalculator />} />
+        <Route path="/employees" element={<EmployeeList />} />
       </Routes>
     </div>
   );
