+++ client/Components/Employee/EmployeeList.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 const EmployeeList = () => {
-  return <div>User List Component</div>;
+  return <div>Employee List Component</div>;
 };
 
 export default EmployeeList;
